remove handles from MoveTo arguments
all calls to MoveTo now should not pass handles as input variables. For
the moment they must all take care of their own display (as in
spectrometer) or the display will not be updated (as in Test Phasing).
In the future the edit box and jog buttons will be part of the motor
class and it will update them directly.