remove errors in pbGo and pbBackground
In pbGo_Callback and pbBackground_Callback no longer rethrow error from
the try catch block. Rather, print the message as a warning, reset the
buttons, the FPAS, and the ScanIsRunning flags. This way problems in the
method or Sampler will no longer crash the whole program. Should be a
little more graceful... Lets see.